ATimer/requirements.txt
flask==2.0.3
Jinja2==3.0.3

pyyaml==6.0
markdown==3.4.1 

protobuf==3.20.1
matplotlib==3.6.2
seaborn==0.12.1
numpy==1.23.5
pandas==1.5.2

pytest==7.2.0